+++ hud.h
@@ -31,7 +31,7 @@ extern int show_coord;
 extern int show_coord_2;
 #endif //SHOW_COORD_SETTER
 extern int rot_boussole; //Type de rotation de la boussole
-extern Uint32 exp_lev[200];
+extern Uint64 exp_lev[200];
 #else //ENGLISH
 /*!
  * \name Quickbar defines
@@ -412,6 +412,6 @@ void change_max_nutri(int max);
 } // extern "C"
 #endif
 
-extern Uint32 exp_lev[200];
+extern Uint64 exp_lev[200];
 
 #endif	//__HUD_H
